requirements.txt
Flask==1.1.1
fasttext
numpy==1.18.3
gensim==3.8.3	
tensorboard==2.4.0
tensorflow==2.3.1
#tensorflow
#tensorflow-estimator
tensorflow-estimator==2.3.0
Keras==2.3.1
xlrd==1.2.0
rasa==2.0.6
#rasa-nlu==0.15.1
openpyxl==3.0.4
sklearn_crfsuite==0.3.6
spacy==2.3.2
absl-py==0.9.0
gast==0.3.3
pandas
inflect
transformers==2.6.0
torch==1.7.1
torchvision==0.8.2

